// interfaces/driverSolicitude.ts
export interface DriverSolicitude 
{
    id: string;
    id_type: string;
    name: string;
    surname: string;
    email: string;
    genre: string;
    phone: string;
    home_address: string;
    city: string;
    country: string;
    birthDate: Date;
    age: number;
    licenseNumber: string;
    licenseType: string;
    licenseDate: Date;
    licenseExpirationDate: Date;
    licenseCountry: string;
}